Hide queue headers for empty queues when filtering
When filtering the status page for Zuul, the queues that don't
match the filter but still have jobs will have their headers
shown, resulting in a long list of nothing.

This patch addresses this issue by making sure that we count
the number of visible change boxes, if there are none, we don't
display the `div` that contains the change queue.

The only caveat of this change is that when hitting 'enter' after
filling out the filter form, it will result in a full reload of
the page and an HTTP request back to the status server, however,
this isn't something that is done often so we can get away
with this until we convert the status page to using Angular
entirely, which will allow for easier filtering.

Change-Id: I939912882ae900ab1fdf1a2adf9f024158572887