Handle more than one websocket streamer at a time
The ws4py websocket manager assumes that the received_message
method will return quickly.  It maintains a single thread for
handling all websockets, but the updated websocket handler in
zuul-web assumed it would run in a thread for each request.

To correct this, start a new single thread in the ZuulWeb object
which handles ongoing finger streaming duties for all websocket
streaming threads.  Whenver a new finger streaming websocket
request is handled, that finger socket is added to a poll list
in the new thread, and any newly received data is forwarded to
its associated websocket.

The test_websocket_streaming unit test is altered to start two
client simultaneously to verify the new behavior.

Change-Id: I505b05b98c2c8fa1b3e875cd5e9266e9929f1367