Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

One minor concern in the testsuite which seems to not currently be a problem based on the +1 from zuul.

Also we might want to amke the websocket code part of its own file for logical separation now that this has become quite a bit more complicated.

Patch-set: 2
Label: Code-Review=+2