Update patch set 2
Patch Set 2: Code-Review-2

I'm leaving a -2 on this for now not because I think we'll never merge it, but because we haven't really discussed whether this is desirable and we have a lot of other things to focus on first.

My first concern: is this useful for the parts of the API which are not in the SQL database?  Most of the interesting part of the API doesn't use the SQL db.  We've also considered storing the build records in something other than a SQL db in the future (maybe zookeeper, maybe something else).

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2