Update patch set 2
Patch Set 2:

I think we need to do something slightly more complex to support all of the use cases. A few thoughts:

/t/tenants.html doesn't make a lot of sense, because /t is the prefix for a specific tenant. (that's not a non-starter - just saying it loos weird)

For whitelabeled deployments, the main link should not be a tenant list. Basically, for multi-tenant, / is tenants.html but for whitelabeled / is status.html. This is why the attempt is to have the navigation bar link be to '/'.

I think we should likely do $something in the routing controller to assign either the status controller or the tenant controller to / (not 100% sure what that is yet)

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0