web: add tenant-scoped, JWT-protected actions
A user with the right JSON Web Token (JWT) can trigger a autohold,
reenqueue or dequeue a buildset from the web API.
These specific endpoints can be deactivated in the configuration.

The Token is expected to include a key called "zuul.tenants" that
contains a list of tenants the user is allowed to perform privileged
actions on. It is up to the operator to provide a way for the user
to obtain such a JWT, for example using pyJWT's CLI:

jwt --key=NoDanaOnlyZuul --alg=HS256 iss=me zuul.tenants=local exp=+3600

The Token can be passed as a bearer token in an Authorization header,
or as the 'jwt' query argument.

Change-Id: Ief9088812f44368f14234ddfa25ba872526b8735