[WIP] web: add tenant-scoped autohold, enqueue
A user with the right JSON Web Token (JWT) can trigger a autohold or
reenqueue a buildset from the web API.
These specific endpoints can be deactivated in the configuration.

The Token is expected to include a key called "zuul.tenants" that
contains a list of tenants the user is allowed to perform privileged
actions on. It is up to the operator to provide a way for the user
to obtain such a JWT.

TODO:

* tests
* doc
* add JWT expiry check to protected()

Change-Id: Ief9088812f44368f14234ddfa25ba872526b8735