web: add tenant and project scoped, JWT-protected actions
A user with the right JSON Web Token (JWT) can trigger a autohold,
reenqueue or dequeue a buildset from the web API.
These specific endpoints can be deactivated in the configuration.

The Token is expected to include a key called "zuul.tenants" that
contains a dictionary of tenants and projects the user is allowed
to perform privileged actions on. It is up to the operator to provide
a way for the user to obtain such a JWT after authentication.

The Token must hold the mandatory claims 'exp' and 'zuul.tenants', for
example:

{
 'exp': 1234567890,
 'zuul.tenants': {'tenant1': ['projectA', 'projectB']}
}

The Token can be passed as a bearer token in an Authorization header,
or as the 'jwt' query argument.

Change-Id: Ief9088812f44368f14234ddfa25ba872526b8735