Update patch set 1
Patch Set 1: Code-Review-1

This was done intentionally (https://review.openstack.org/433748) as old versions of Gerrit return the change number as str and new ones as int so this is normalized to str. In case there are problems with refresh_deps we probably have a further occurrence somewhere which needs to be normalized to str and I'd prefer to fix that instead.

Patch-set: 1
Label: Code-Review=-1