Fix missing dequeue on synchronized pull request variant one
When updating a pull request Github can automatically dismiss a
review. In this case Github also sends a pull_request_review
event. This event contains the old head of the pull request which
leads to zuul updating its internal state of the old pull
request. This also updates the updated_at timestamp. Unlike Gerrit
there is no notion of a by time ordering of pull request 'versions' so
we currently rely on the updated_at timestamp to check if a pull
request is newer than another one. So in cases like mentioned above
that fails because the old and new pull request object finally get the
same timestamp which prevents dequeueing of the old pull request.

As the updated_at timestamp is a moving target we cannot rely on that
to determine if a pull request is newer than another one. Under the
assumption that we check this only for newer pull requests based on
the event inflow we can just ignore the timestamp and dequeueing
works.

Change-Id: I1d697aadf0e0c716117396f97189a4a77fc56ca8