Update patch set 9
Patch Set 9: Code-Review+2

So this change looks fine to me and is probably the simplest approach.

However, it doesn't feel right having the main() method called and then exited from early. The parameter is really changing the behavour of the command that the rest of the set up doesn't make sense. I wonder if we should implement this as a separate sub-command or route it elsehow through the app instead of main(). I realise this gets convoluted with the socket commands passed to a running scheduler though so it's not a straight forward change.

Patch-set: 9
Label: Code-Review=+2