Update patch set 1
Patch Set 1: Code-Review-1

Just tested against our github and it indeed returns 404 if branch protection is disabled (and the branch exists). However Joshua is right, we should test this. The test infrastructure for this is there (FakeGithub class) and probably only needs to be extended (or fixed) slightly.

Patch-set: 1
Label: Code-Review=-1