Update patch set 1
Patch Set 1:

There are two ways in k8s to configure applications. One is env vars, the second is mounted configmaps and secrets. The most common is env vars which is recommended generally (but in my eyes this makes only sense if it's only env vars). In zuul we need config files so at least I personally think it should be either config files or env vars.

A different way (which also enables more use cases) to solve that problem is to support a zuul.d directory and let zuul merge its configuration from multiple files. That way the secrets and rest of configs can be mounted as different files. That's what we have done in form of a pre run script. But I'd prefer to support that natively in zuul in the long run.

Patch-set: 1