Update patch set 1
Patch Set 1:

As Tobias already mentioned, keeping secrets in ENV is a common and recommended practice in containers. There it has no security concern, since only one app is running there. Having secrets mounted as files as well and being merged by zuul should be also nice option if ENV is still not wanted (agree, that it should not be used in normal installation). The reason for this change initially, is that some secrets (GitHub, DB) should not be stored simply unencrypted in a configMap, which serves as base for zuul.conf, since ConfigMap can be easily retrieved by anyone, who has access to k8/Openshift zuul namespace

Patch-set: 1