Update bindep file with build profiles
Looking ahead towards potentially using pbrx to build containers of
zuul, we want to indicate some bindep packages are needed only for
building a wheel, not for installing/running it.

Introduce a 'build' profile that can be used to get the distro dependencies
needed for building wheels of the python dependencies. Also add runtime
library depends without a profile so that bindep -b properly reports the
packages needed for installing if wheels have been built.

Zookeeperd is moved to the 'test' profile as it is not required to be
installed locally for zuul to operate.

Depends-On: https://review.openstack.org/580521
Change-Id: I3c8dcedddd8634661b3524966a6f228c9edde29b