Build container images using pbrx
The pbrx project has a command to build container images based on the
python:alpine base image for any project using pbr and bindep.

Use it to start building container images for zuul. A followup patch
will add support for publishing the images to dockerhub.

The images want the database drivers in them to be useful out of the
box. Also, the executor wants ara installed. By adding those to extra
entries named after the resulting image name, pbrx will install the
extra packages at the right time, as well as produce wheels for them
during the wheel cache stage.

Depends-On: https://review.openstack.org/582380
Change-Id: I453b0ab2c4916cb77249b5f1039e6e18fde47ab1