Update patch set 1
Patch Set 1: Code-Review-1

I think in a matches* function it is bad to return True or a string as both evaluate to True which is definitely a surprising behavior of such a function. I agree that it is useful to have more information in the response but we should find a way to retain the automatic evaluation to False when this function is called without caring about the reason.

Patch-set: 1
Reviewer: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1