Reduce number of reconfigurations on branch delete
If a branch is deleted that didn't contain configuration files at all,
we do not need to trigger a tenant reconfiguration. Since we can only
check the unparsed configuration cache during a reconfiguration, we need
to add the information about the branches affected and the action taken
on these branches to the event. As a by-product, we can then also reduce
the number of fetched configurations per repository by deleting
configurations on a pre-branch basis.

Change-Id: I34f5d3631d8112c792a72146268dad85813ceb4f