Update patch set 4
Patch Set 4:

Re: s2i - I've looked at it a bit and it's promising, but it's still a decent amount of work to get the results we'd want. That said - at least some of what's in pbrx could be implemented to use s2i under the covers.

The main difference is that s2i is a general tool, and thus flexible, which means harder to use. What pbrx is trying to do is make things easy by removing flexibility - which is fine in the context of a project that is using pbr and bindep.

The part of pbrx that s2i does very well is the multi-stage build - where we're using a builder image to make wheels and then putting those wheels into the final base image.

What it's not suited for (and I don't blame them) is figuring out a list of images and entrypoints from the python setup.cfg console_scripts metadata.

All that is to say- I agree, we should keep s2i on our radar and should definitely noodle with it some. It has several things going for it - and if we can use it with good results that's potentially better than keeping our own tool.

Patch-set: 4