Update patch set 1
Patch Set 1: Code-Review-1

This only returns the parent job of the *last* variant of a given job.  There's a fundamental data modeling problem here -- each job variant can have a different parent.  This is the wrong place to provide the parent info, it should be on the particular job variant (and it already is).

I think our time will be better spent focusing on individual jobs, rather than creating a graph of all the jobs in the system.  It might look cool (until it gets large enough), but it's going to be more useful for a user to be able to look at a particular project or job and see the relationship there.  Looking at a job and seeing the entire parent/child hierarchy from that job -- that's going to be great.

Patch-set: 1
Label: Code-Review=-1