timer: skip projects not using the pipeline
This change makes the timer driver skip projects that are not configured
to use the timer triggered pipeline. Timer operates on branch heads and
doesn't need speculative layouts to decide if it should be enqueued or not.
So it can be decided on cached data if it needs to run or not.

Change-Id: I410c0da5057727734ac82ce56cad75655a9ebc40