Update patch set 1
Patch Set 1:

> This only returns the parent job of the *last* variant of a given
 > job.  There's a fundamental data modeling problem here -- each job
 > variant can have a different parent.  This is the wrong place to
 > provide the parent info, it should be on the particular job variant
 > (and it already is).
 > 

This enables efficient jobs grouping and it seems like the right place to provide the parent info. The issue is that we only return the last variant of a given job, and maybe we should extend the interface so that each variants list can be queried. E.g.: api/jobs/master   and api/jobs/stable/queens.

Patch-set: 1
Label: Verified=0