Move the database libraries to requirements
In order to make container images that are generally applicable and
usable by people, we need to have database drivers in place. Currently
we only list the libraries in test-requirements. However, PyMySQL is a
pure-python library so should not be an undue burden on people not using
it. Similarly, psycopg2-binary has wheels published for all the
platforms people normally use, so should not generally be an issue either.

psycopg2-binary does not have alpine wheels, so on alpine this will
cause posgresql-dev to be needed for installation. We have a solution
coming using builder images and wheel transfer from builder image to
final image that should mitigate this.

Change-Id: I004ffc52fb2a57cb7f5ea6bd30c49afa766ba5f2