Add tenant yaml validation option to zuul client
This patch adds a new command 'tenant-conf-check' to the Zuul client
command. This option runs a tenant_file validation by running the schema
valiation of the file. The command exits -1 if errors have been
detected. The command does not use RPC call but instead expects to
find the tenant_file on the local filesystem.

Change-Id: I6582bbc37706971085dac5c3ca3b4c690c515f9e