Update patch set 10
Patch Set 10:

> So this change looks fine to me and is probably the simplest
 > approach.
 > 
 > However, it doesn't feel right having the main() method called and
 > then exited from early. The parameter is really changing the
 > behavour of the command that the rest of the set up doesn't make
 > sense. I wonder if we should implement this as a separate
 > sub-command or route it elsehow through the app instead of main().
 > I realise this gets convoluted with the socket commands passed to a
 > running scheduler though so it's not a straight forward change.

Yes that's true, so I decided in the PS10 to move the code in the zuul client tool. I hesitated to create a new command under cmd/ but it was far simpler to put it in the client. Hope that's better like that. Thanks for the review Joshua.

Patch-set: 10
Label: Verified=0