Update patch set 1
Patch Set 1:

We probably want release notes and docs for this (including, perhaps adding a deprecation note for the signal-based full reconfiguration).  They don't have to be in this change, we can do that in a later change, but we shouldn't merge this one until it's been written.

Otherwise, lgtm.

Patch-set: 1
Label: Verified=0