Optionally support mitogen for job execution
Using Mitogen [1] when running Ansible can have potentially
substantial performance improvements and reduction of resource
usage. Especially on larger deployments this may have a noticeable
impact on the executor performance. Thus add it as an optional feature
that can be enabled on a job so we can try it in real life. We don't
know yet if this works for any job so keep it disabled by default for
now and mark it as experimental in the docs.

[1] https://networkgenomics.com/ansible/

Change-Id: I9333db4de989123daee4da760a56ae7820d2d444