Per-branch management of unparsed config cache
Currently the unparsed config cache gets erased if a project triggers a
reconfiguration because one branch was changed or created in that
project. This patch adds infrastructure to manage the cache on a
per-branch basis by adding the branch(es) to the tenant reconfigure
event. This will save some configuration fetches and reduce the load on
SCM.

Change-Id: I023d9f88f478a21aca4f121f4e2c89d6ff9bb091