Update patch set 11
Patch Set 11:

> Hm, I'm just thinking if that should be part of the zuul client.
 > The zuul client typically communicates with the scheduler via rpc.
 > The check-tenant-config is an offline check that might be better
 > suited in the zuul-executor script. How do you think about this?

My concern if we put it in the executor is that the tenant_file is not guaranteed to be on the local FS. But after all, this tenant-conf-check  is suited to verify tenant configuration prior to its installation, so with a fake zuul/config.conf. Within Software Factory we do this validation on the executor nodes so well it makes sense to put it in the executor too. I'll do a follow up change to move that in the executor. Thanks for the review !

Patch-set: 11