Update patch set 12
Patch Set 12: Code-Review-1

Since this doesn't rely on a running Zuul at all, it should be pretty easy to write a test for it (just exec the command with a working and non-working tenant yaml).  I think we should do that; otherwise it is certain to break.

Patch-set: 12
Label: Code-Review=-1