Update patch set 1
Patch Set 1:

This approach seems fine, though the arbitrary cache size perturbs me a bit.

We could, instead of this, simply decrypt the secrets once and store the result on the secret object itself.  Currently we store the secret encrypted in the layout and configuration cache, and only decrypt it when freezing the job as extra protection against an accidental coding error which used a decrypted secret in a different project.  However, we could take the appropriate amount of care when using the secret object to make sure that it isn't used in the wrong project.  That would probably result in simpler secret code which should be easier to follow, and would eliminate the need for an external cache.

Monty has suggested that we consider allowing secrets to be used across projects in some specific circumstances (eg, the "provide a secret as an argument to a parent job" use case).  We haven't thought about that in much detail yet, but if we do that, we may end up reworking some of the secret code in a way which eliminates some of the protections we currently have anyway.

Patch-set: 1