Update patch set 20
Patch Set 20: Code-Review-1

(3 comments)

-1 for --tenant argument I think we want to move it

As for other comments, I think there is an optimization here when using --change, I believe the project is already included in that info, so passing --project seems extra.

We also don't appear to be testing --ref, think we could also add a test for it.

Patch-set: 20
Label: Code-Review=-1