Update patch set 2
Patch Set 2: Code-Review-1

I dig the idea of project variables, but I think these should be be merged in with the regular job variables, not put in the special hierarchy for zuul variables.  "zuul." is for variables that originate from zuul itself (information about the node, the executor, the change, etc).  Job variables don't have any particular prefix, with the idea that they are variables that you would set and use even if you were running a playbook outside of Zuul.

Here you can see the existing types of variables and their order of precedence:
https://zuul-ci.org/docs/zuul/user/jobs.html#variables

I would put these between "job variables" and "parent job results"; ie, the second-lowest order of precedence, so that if you set a project variable *and* a job variable with the same name, the job variable would win.

Patch-set: 2
Label: Code-Review=-1