Support job pause
There are some use cases where it would be very useful to have some
resources that stay during the livetime of the buildset. E.g.:

* Spinning up a docker registry for child jobs that automatically gets
  cleaned up after a test run.

* Efficiently caching intermediate build results that are needed in
  child jobs without needing and overloading a central system for
  this.

In order to provide this functionality a job now can pause itself
after the run phase and let its children run. After all children are
finished it will be resumed. This can be easily triggered in a job by
using the zuul_return module together with the ability to forward any
data via ansible variables to the children.

 - name: Pause job
   zuul_return:
     data:
       zuul:
         pause: true
       registry_ip: "{{ hostvars[groups.all[0]].ansible_host }}"

Change-Id: I1a078bf22b9d3f9a33bd1517e73234749932e1bf