Rename didAnyJobFail() to hasAnyJobFailed()
I feel these functions are slightly confusingly named as is, because
you would expect that

 didAnyJobFail() == not didAllJobsSucceed()

but that isn't really the case.  didAnyJobFail is being used on the
active job list and is ignoring any running jobs without a return
status yet.  I think it's better named as a present-tense
"hasAnyJobFailed".

didAllJobsSucceed() is being used to check the return status of all
jobs, and gives an error if a job doesn't have a build result.  One
slight enhancement to give the "skipped" variable a more
intrinsitically obvious name.

Documentation is added for both functions.

Change-Id: Ic19f929259cd9894146b128cf9bc84e439104671