Update patch set 4
Patch Set 4: Code-Review-1

I love this patch and the impl looks good to me. There is one thing I'm unclear on, and that's precedence between extra-vars and site-variables (and it doesn't look like the test validates that)

I'd expect site-variables to win in all cases. Is that what's happening?

I'm only leaving the -1 for that question - if we're good, then +2 (but also it needs a rebase)

Patch-set: 4