Update patch set 1
Patch Set 1:

There's no "right" size for this cache (that's the core of my objection), so I don't think you should change it.

If we cache the decrypted versions of the objects as I suggested, we would only decrypt them again on changes to the repo in which they are stored, just like parsing any other config object.

I could see keeping a smaller cache in that case (so that if, say, a particular repo with a lot of secrets merges a lot of changes in a time period), that case is sped up a bit.  I like that idea.

Patch-set: 1