WIP: Unexpected behavior of a reject clause
This change demonstrates a behavior that might be surprising to a
user: If a reject clause is configured for a pipeline, the reject clause
is always true even if no approval is set. If, however, an unrelated
approval is added to the change, the reject clause is evaluated as
expected.

We should discuss whether this behavior should be changed or at least
documented since it seems surprising that a reject always hits, even in
the absence of the referenced approval, but only if no other approval is
present.

Change-Id: I220bbed931ced7e807febefbf2b1b33af53a8da0