WIP Map file comment line numbers (alternate)
This is an alternative sketch to https://review.openstack.org/590429

It will be faster and more efficient, but if the job has altered
the git repo, it may fail or produce incorrect data.  We could accept
that as a documented caveat.  It doesn't seem especially likely
to happen for the sort of jobs which would produce line comments.

Change-Id: Iad48168d41df034f575b66976744dbe94ec289bc