Update patch set 2
Patch Set 2:

I think there is also a thing - that I ran in to when thinking about using making add_host usable by end users with bridge.openstack.org ... and that is ssh keys.

We don't have a good way for a user to add a private key to the existing ssh agent. We can put a private key in a secret, but getting the ansible run by zuul to use that private key is a whole other issue.

If we do something like this (which is sort of where my brain started going) then the user could do this and it would work.

Granted - this is basically re-impl-ing add_host in some ways.

SO - maybe we also want to consider writing our own add_host override module that behaves like add_host but also spits out zuul_return info at the end to inject the host into the inventory?

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0