Refresh job page on navigation
By default, Angular doesn't refresh a page if you navigate to it
(ie, click a link that points to it) while you are on that page.
This means that if even the query string changes, the change won't
update.  You can, however, tell it to fire navigation events when
the query string changes, and then hook into those events to update
the page.  This change does that.

This causes the jobs.html page to navigate "normally" (ie, the back
button works) when clicking on a parent job link.

This also adds the "Branches" attribute to the table.

Depends-On: https://review.openstack.org/591089
Change-Id: I7397bda1e841916a69789b1116c40256a198edb7