Update patch set 4
Patch Set 4: Code-Review-1

(7 comments)

I'm still reading/learning - these are initial comments.

Overall I agree (somewhat surprisingly) that the react jsx code is easier to read/follow. I'd suggest that we shift the file extensions to .jsx instead of .js just so that it's easier for editors and whatnot to know how to syntax highlight it.

I don't think we're likely to split zuul-web out in the near term - especially while we're still working heavily on the REST API. That said - I don't mind pushing everything down into the web directory like this.

I mentioned it inline, but it's important to keep the existing status page design. Let's check in with other people on viability of this work looking at the jobs / builds pages first before spending the effort to wire that in. I'm also curious to see how the websocket streaming works with the react tools - I'd been poking at trying to make the angular streaming code more event-driven. But again, before spending effort on that let's see what people think about the way these pages work.

It would also be a good idea to explore BrowserRouter and see what that looks like.

To me, this is good enough to bother other people for their opinion.

Patch-set: 4
Label: Code-Review=-1