[WIP] Fix post pipeline change display
The abbr tag isn't working for some reason (it seems to be stripped from
the dom I see in the debugger in firefox) It's a nice idea, but for now
just remove it and revert back to the behavior from the jquery version
of just sticking the changeTitle in there when it exists. Otherwise
we're printing 40-char uuids and blowing the table.

Change-Id: I3183e8ab4c1c13ef9fdf8c9ec573b56608503e87