Update patch set 1
Patch Set 1:

I'm not opposed to adding more queryable parameters, but I don't understand the commit message.  The unique identifier for a build is, in fact, the uuid, which is already queryable.  I think it's fine to add this in case someone wants to copy/paste a log url into a search field to find a build.  But I don't think we should use this for internal build queries (ie, as the basis for a 'build' page).  For that, we should use https://review.openstack.org/592225 which you may want to review.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0