Update patch set 24
Patch Set 24:

Starting with PS23, I've re-organized the code and changed the logic a bit to remove the global store constant. Pages that need tenant information gets it through the tenant property set by the Redux connection.

This mostly enables a cleaner interface for tests, and I've added one for the ChangePanel to test links for multi-tenant and white-label mode.

Patch-set: 24