Reload tenant in case of new project branches
Since GitHub doesn't report branch protection changes, it is possible
that a pull request to a branch is not considered since the
configuration for that branch isn't loaded yet. This is, however,
surprising for the user.

This change updates the branch protection cache if a pull request was
received and when the trigger event is then proecessed in the scheduler,
it checks if the target branch is a protected branch whose config is not
yet in the tenant's configuration cache. If that's the case, a tenant
reload is issued.

Change-Id: I0d851a8773b73de54b0b777691a6a0d9f788f096