Update patch set 4
Patch Set 4:

> I think we need to discuss this a bit before we embark on this.  We
 > chose statsd because it is a very simple and low-impact method of
 > reporting metrics.  It incurs almost no overhead in the application
 > itself which allows us to freely add metrics to any activity we
 > like without worrying about impact to the application.  It has a
 > simple wire protocol for which anyone can implement backends, and
 > indeed, there are many backends for it already.
 > 
 > Including prometheus.
 > 

Prometheus also has a simple wire protocol, you can query the metrics with curl /metrics, it's text based.


 > Prometheus, on the other hand, has a very heavy-weight architecture
 > which imposes a significant resource load on the application
 > itself.  Statsd's best-effort publishing approach has a number of
 > advantages over prometheus's polling.
 > 

I think it depends how this is measured. Statsd traffic increases with the volume of metrics while prometheus poll them all in one connection...


 > I think Zuul should only have one metric system, and I think statsd
 > is the better choice due to the simplified application requirements
 > and near-standard protocol with varied backends.

The goal is here is to enable executor auto-scale based on job queue and be able to send alert if enqueue time goes above a limit. To set this up, all you need is a single prometheus service to poll the scheduler and a couple of alert configured. What would be the procedure to enable those 2 simple things with statsd?

Another use-case is to get an alert if a service goes down, which happens by default with prometheus since it polls the service.

Patch-set: 4