Update patch set 6
Patch Set 6: Code-Review-1

I think we should have a test with a parent job with a single playbook and a child job with two playbooks where we assert that the parent job's playbook did not run but both child playbooks do.

Basically, I want to codify in tests that the behavior for inheritance of 'run' is to override, rather than append or prepend for pre-run and post-run.

Patch-set: 6
Label: Code-Review=-1