Update patch set 4
Patch Set 4:

Tristan Cacqueray wrote:

> Prometheus also has a simple wire protocol, you can query the
> metrics with curl /metrics, it's text based.

HTTP is not a simple wire protocol.  My point was that it is only a few bytes of network traffic and a few CPU cycles to emit a metric with statsd.  Prometheus is far more heavyweight no matter how you measure it.

>  > Prometheus, on the other hand, has a very heavy-weight
> architecture  > which imposes a significant resource load on the
> application  > itself.  Statsd's best-effort publishing approach has
> a number of  > advantages over prometheus's polling.  >
>
> I think it depends how this is measured. Statsd traffic increases
> with the volume of metrics while prometheus poll them all in one
> connection...

At the cost of accumulating the metrics in memory on the server.  With prometheus, the resolution must decrease in order to accommodate that.  Statsd can emit data instantaneously at almost no cost.

If needed, we can, of course, batch or roll-up stats in statsd.  Ian wrote some changes to bath reports for better network performance

> The goal is here is to enable executor auto-scale based on job queue
> and be able to send alert if enqueue time goes above a limit. To set
> this up, all you need is a single prometheus service to poll the
> scheduler and a couple of alert configured. What would be the
> procedure to enable those 2 simple things with statsd?
>
> Another use-case is to get an alert if a service goes down, which
> happens by default with prometheus since it polls the service.

Okay, that's something that prometheus is better at.  How about we consider a limited use of prometheus to enable the auto-scale functionality while retaining statsd for general metrics?

I think the best metric for auto-scaling executors would be related to the executor queue length.  Do you have an idea for how to use that to determine when a new executor was needed?

Patch-set: 4