Update patch set 4
Patch Set 4:

If the overhead is a concern, then it seems like using prometheus for a couple of metrics will have the same cost as for many metrics... However, afaict, the overhead is negligible: the default scrap interval is only every 15 seconds and on the service side, updating a metrics is mostly in-memory update.

The merger_queue and executor_queue gauge defined in https://review.openstack.org/#/c/599210/3/zuul/scheduler.py should be enough to trigger scale-up/scale-down action.

Another thing to consider is that grafana + prometheus is relateively easier to setup than graphite/whisper/statsd.

Patch-set: 4